STYLE: Remove ad-hoc code patching.
The ad-hoc code patching by string replacement is not a sustainable,
portable, or maintainable solution to making VTK build.  It conflicts
with other necessary patch sets and rapidly becomes a tangled mess.

Please use a private git repository with patches that are intended
to be contributed back to the main tree when necessary.